Remove the overlay logout step
A view which returns the logout url is used, since the url is generated
dynamically.

Since the logout overlay (now logout url) is now generic, it is moved to
opengever.base

Also the CAS test case remains functional in order to pass.
