+++ opengever/base/browser/configure.zcml
@@ -224,4 +224,11 @@
       permission="cmf.ManagePortal"
       />
 
+  <browser:page
+      for="*"
+      name="logout_url"
+      class=".logout_url.LogoutURL"
+      permission="zope2.View"
+      />
+
 </configure>
