+++ opengever/core/upgrades/20180711170915_remove_logout_overlay/upgrade.py
@@ -1,0 +1,9 @@
+from ftw.upgrade import UpgradeStep
+
+
+class RemoveLogoutOverlay(UpgradeStep):
+    """Remove logout overlay.
+    """
+
+    def __call__(self):
+        self.install_upgrade_profile()
