--- opengever/document/browser/templates/logout_overlay.pt
@@ -1,21 +1,0 @@
-<form name="logout_overlay" method="POST" action="./logout_overlay" i18n:domain="opengever.document">
-
-      <h1 class="documentFirstHeading"
-          i18n:translate="alert_not_checked_in_documents">
-        You have not checked in documents. Do you realy want to logout?
-      </h1>
-
-      <h2 i18n:translate="overview_not_checked_in_documents">
-          Following documents are checked out:
-      </h2>
-
-      <ul class="subContents">
-        <li tal:repeat="item view/items">
-          <a target="_blank" tal:attributes="href item/getURL" tal:content="item/pretty_title_or_id"/>
-        </li>
-      </ul>
-      <input type="hidden" class="logout_overlay_redirect" name="form.redirect.url"
-             tal:attributes="value view/redirect_url" />
-      <input type="submit" name="form.submitted"
-             value="Logout" class=" logout_overlay_submit destructive" i18n:attributes="value"/>
-</form>
